+++ laktory/models/__init__.py
@@ -3,8 +3,10 @@ from .column import Column
 from .database import Database
 from .dataevent import DataEvent
 from .dataeventheader import DataEventHeader
+from .datasources import *
+from .grants import *
 from .pipeline import Pipeline
 from .producer import Producer
-from .datasources import EventDataSource
-from .datasources import TableDataSource
 from .table import Table
+from .users import *
+from .volume import Volume