+++ laktory/models/database.py
@@ -3,6 +3,7 @@ from pydantic import computed_field
 
 from laktory.models.base import BaseModel
 from laktory.models.table import Table
+from laktory.models.grants.databasegrant import DatabaseGrant
 
 
 class Database(BaseModel):
@@ -10,6 +11,7 @@ class Database(BaseModel):
     comment: Union[str, None] = None
     tables: list[Table] = []
     catalog_name: Union[str, None] = None
+    grants: list[DatabaseGrant] = None
 
     # ----------------------------------------------------------------------- #
     # Computed fields                                                         #