+++ laktory/models/users/__init__.py
@@ -1,0 +1,4 @@
+from laktory.models.users.user import User
+from laktory.models.users.usersgroup import UsersGroup
+from laktory.models.users.serviceprincipal import ServicePrincipal
+